(1767) séparation barre de recherche annuaire en composant séparé
