+++ packages/frontend/webapp/src/views/CommunauteView.vue
@@ -14,8 +14,7 @@
 </template>
 
 <script setup>
-import { computed, onMounted } from "vue";
-import router from "@/helpers/router";
+import { onMounted } from "vue";
 import { useQuestionsStore } from "@/stores/questions.store";
 import { useDirectoryStore } from "@/stores/directory.store";
 
@@ -25,31 +24,6 @@ import Annuaire from "@/components/Annuaire/Annuaire.vue";
 
 const questionsStore = useQuestionsStore();
 const directoryStore = useDirectoryStore();
-
-const search = computed({
-    get() {
-        return {
-            search: "",
-            data: null,
-        };
-    },
-    set(newValue) {
-        if (newValue) {
-            if (newValue.data?.type === "user") {
-                router.push(`/annuaire/${newValue.data.organization_id}`); // TODO
-            } else if (newValue.data?.type === "organization") {
-                router.push(`/structure/${newValue.data.id}`);
-            } else {
-                // location ou recherche textuelle
-                directoryStore.filters.search = newValue.search;
-                directoryStore.filters.location = newValue.data;
-            }
-        } else {
-            directoryStore.filters.search = "";
-            directoryStore.filters.location = null;
-        }
-    },
-});
 
 onMounted(() => {
     if (questionsStore.questions.length === 0 || !directoryStore.isLoaded) {
